requirements.txt
fastapi
uvicorn
requests
PyJWT
cryptography
